Rename keeperImporters to keeper
Rename the folder
Fix all the imports
Remove keeperImporters from white-list
